refactor : Header.server > Header로 수정
